Started to implement open end times.
It would be nice as noticed by AMDmi3 to have "something like handling open end
as a zero-length interval (10:00-10:00) with a way to check that the closing
time has a special meaning."

As we now have the state unknown and comments I guess we could use this for
open end times if there are no comments specified by the mapper. But there is
one problem … One rule can only have one state and one comment. This needs to be
fixed first.