Relicensed opening_hours.js under LGPL-3.0 only. Related to #133.
This change applies to all modifications made by me to the opening_hours.js file.

# Reason for LGPL-3.0

The main reason for the license switch is to ensure that all modifications to this library will be required to stay free and be published in source code under the same licenses (LGPL-3.0) while still allowing for the license to be used in projects not licensed itself under a strong copyleft license*. This is expected to best suite the free Software community.

* Such as the GPL or AGPL.

This change was also made as preparatory work for the rewrite of this library in the Haxe programming language (tracked in https://github.com/opening-hours/opening_hours.js/issues/136) because Haxe is able to easily target compiling languages which could be used to distributed this library in binary form only.

# LGPLv2.1?

* Pro: Allows usage in projects which are GPL-2.0 only (https://www.gnu.org/licenses/gpl-faq.html#AllCompatibility)
* Contra: Dual licensing would mean that restrictions/rework done in (L)GPL-3.0 would not apply because someone using the library could decide to use it under the LGPL-2.0 only. https://www.gnu.org/licenses/rms-why-gplv3.html

Decision: Only license under LGPL-3.0. Sorry for anyone using GPL-2.0 only. You might consider updating or using this library in a way which does not violate the LGPL-3.0.
I am not against using this library in your GPL-2.0 only licensed project.

# Reasons for not including the 'or (at your option) any later version [of the License]'

Updated versions of the (L)GPL will need to be evaluated manually by the maintainer of opening_hours.js and approved before licensing this library under future versions.